Refactor to use date objects instead of formatted strings
