--- src/routes/utils/api.ts
@@ -1,53 +1,0 @@
-import { getQuery } from 'api/queries';
-import type { AccountSummaryReport } from 'api/reports/accountSummaryReports';
-import { ReportPathsType, ReportType } from 'api/reports/report';
-import type { AxiosError } from 'axios';
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import type { AnyAction } from 'redux';
-import type { ThunkDispatch } from 'redux-thunk';
-import type { RootState } from 'store';
-import { FetchStatus } from 'store/common';
-import { reportActions, reportSelectors } from 'store/reports';
-
-interface AccountSummaryStateProps {
-  summary?: AccountSummaryReport;
-  summaryError?: AxiosError;
-  summaryFetchStatus?: FetchStatus;
-  summaryQueryString?: string;
-}
-
-export const accountSummaryMapToProps = (deps = []): AccountSummaryStateProps => {
-  const dispatch: ThunkDispatch<RootState, any, AnyAction> = useDispatch();
-
-  const query = {
-    // TBD...
-  };
-  const summaryQueryString = getQuery(query);
-
-  const reportPathsType = ReportPathsType.accountSummary;
-  const reportType = ReportType.billing;
-
-  const summary = useSelector((state: RootState) =>
-    reportSelectors.selectReport(state, reportPathsType, reportType, summaryQueryString)
-  );
-  const summaryError = useSelector((state: RootState) =>
-    reportSelectors.selectReportError(state, reportPathsType, reportType, summaryQueryString)
-  );
-  const summaryFetchStatus = useSelector((state: RootState) =>
-    reportSelectors.selectReportFetchStatus(state, reportPathsType, reportType, summaryQueryString)
-  );
-
-  useEffect(() => {
-    if (summaryFetchStatus !== FetchStatus.inProgress) {
-      dispatch(reportActions.fetchReport(reportPathsType, reportType, summaryQueryString));
-    }
-  }, deps);
-
-  return {
-    summary,
-    summaryError,
-    summaryFetchStatus,
-    summaryQueryString,
-  };
-};
